# api/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Go Warehouse System API
  description: HTTP API for authentication and order creation in the Go Warehouse System.
  version: 1.0.0

servers:
  - url: http://localhost:8081
    description: Auth service
  - url: http://localhost:8082
    description: Order service

tags:
  - name: Auth
    description: User authentication and token management.
  - name: Orders
    description: Order creation.

paths:
  /api/v1/auth/register:
    post:
      tags:
        - Auth
      summary: Register a user
      description: Creates a new user and returns an access token and refresh token.
      operationId: registerUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserRequest"
            examples:
              default:
                value:
                  email: user@example.com
                  password: password123
      responses:
        "201":
          description: User registered successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenPair"
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          $ref: "#/components/responses/Conflict"
        "500":
          $ref: "#/components/responses/InternalServerError"

  /api/v1/auth/login:
    post:
      tags:
        - Auth
      summary: Log in
      description: Authenticates a user and returns an access token and refresh token.
      operationId: loginUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserRequest"
            examples:
              default:
                value:
                  email: user@example.com
                  password: password123
      responses:
        "200":
          description: User logged in successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenPair"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalServerError"

  /api/v1/auth/refresh:
    post:
      tags:
        - Auth
      summary: Refresh tokens
      description: Revokes the current refresh token and returns a new token pair.
      operationId: refreshToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshTokenRequest"
            examples:
              default:
                value:
                  user_id: 8b8a8b1e-0d7f-4f0e-9d88-1a2b3c4d5e6f
                  raw_refresh_token: refresh-token-value
      responses:
        "200":
          description: Tokens refreshed successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenPair"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalServerError"

  /api/v1/orders/:
    post:
      tags:
        - Orders
      summary: Create an order
      description: Creates an order for the authenticated user and reserves stock through the warehouse service.
      operationId: createOrder
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderHttpRequest"
            examples:
              default:
                value:
                  items:
                    - sku: SKU-001
                      quantity: 2
                    - sku: SKU-002
                      quantity: 1
      responses:
        "201":
          description: Order created successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalServerError"

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  responses:
    BadRequest:
      description: Invalid request body or parameters.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Missing, invalid, or expired credentials.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Conflict:
      description: Resource conflict.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    InternalServerError:
      description: Internal server error.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"

  schemas:
    CreateUserRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          example: user@example.com
        password:
          type: string
          minLength: 8
          example: password123

    RefreshTokenRequest:
      type: object
      required:
        - user_id
        - raw_refresh_token
      properties:
        user_id:
          type: string
          example: 8b8a8b1e-0d7f-4f0e-9d88-1a2b3c4d5e6f
        raw_refresh_token:
          type: string
          example: refresh-token-value

    TokenPair:
      type: object
      description: Current Go response uses exported field names because TokenPair has no json tags.
      required:
        - AccessToken
        - RefreshToken
      properties:
        AccessToken:
          type: string
          description: JWT access token.
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        RefreshToken:
          type: string
          description: Raw refresh token.
          example: 1f2e3d4c5b6a7980

    CreateOrderHttpRequest:
      type: object
      required:
        - items
      properties:
        items:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/CreateOrderItemRequest"

    CreateOrderItemRequest:
      type: object
      required:
        - sku
        - quantity
      properties:
        sku:
          type: string
          example: SKU-001
        quantity:
          type: integer
          format: int64
          minimum: 1
          example: 2

    Order:
      type: object
      description: Current Go response uses exported field names because Order has only db tags.
      required:
        - ID
        - UserID
        - Status
        - CreatedAt
      properties:
        ID:
          type: string
          example: 9f9a9b1e-0d7f-4f0e-9d88-1a2b3c4d5e6f
        UserID:
          type: string
          example: 8b8a8b1e-0d7f-4f0e-9d88-1a2b3c4d5e6f
        Status:
          $ref: "#/components/schemas/OrderStatus"
        CreatedAt:
          type: string
          format: date-time
          example: "2026-08-14T10:30:00Z"

    OrderStatus:
      type: string
      enum:
        - pending
        - placed
        - failed
      example: placed

    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          $ref: "#/components/schemas/ErrorBody"

    ErrorBody:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: bad_request
        message:
          type: string
          example: invalid request